More performant implementation of read status for entire forum.
The new implemention doesn’t query each single topic on itself,
but rather retrieves the list of read topics by this user and
compares its length to the total number of topics in the forum.
This results in constantly two queries per forum regardless of
the number of topics in the forum.