Use <div> for CodeRay styling and prevent spoilers.
Code is not tabular material, so <table> is semantically wrong.
Also, it is impossible to style kramdown’s CodeRay output to
overflow with "auto" because it wraps the <table> inside a
<div> element without any class or id attribute.

Line numbers will not be copied on selection as this is prevented
by CSS (at least for Firefox and Webkit-based browsers).