Cleanup for restart.
As a competitor to phpBB and FluxBB, nobody would ever use this
software probably, me included. Even though it was clean code.

Instead, I am going to refocus on different things. For the
Secretchronicles project[1] I need a forum software that seemlessly
integrates with mailinglists. None of the existing forums can do that
properly, and Chessboard's mailinglist plugin was rather
buggy. Setting the focus now onto this very topic and leaving the rest
to phpBB and FluxBB should allow this software to enter into a niche
between forum and mailinglist software.

[1]: http://www.secretchronicles.de