Add option to remove [ml-tag].
Many mailinglist have a bracketed tag in their title indicating their
name (though this is probably obsoleted by introduction of the List-*
headers, but many people still like it). Since these look weird on a
forum, this commit gives the possibility to remove them by setting the
tag's content on the forum instance.