Extract ML monitor into separate daemon.
This setup is much cleaner compared to the hacky at_exit approach that
was doomed to fail when Chessboard is run over multiple processes for
load balancing. This comes at the cost of a little more administrative
overhead, though, as the mailinglist monitoring daemon now has to be
started separately from the Chessboard web application itself.