Keep track of user's read posts.
This commit introduces the required posts-users crosstable for keeping
track of the posts a user read. Each time he visits a thread or topic,
all posts shown to him are marked as read (in the topics view this
properly only marks the posts on the current page as read, in the thread
view the entire thread is marked read). The forum list only checks if
the last post in a thread was read, though, as this makes things easier,
more performant, and will usually deliver the correct result.