Keep track of users switching their display name.
I.e. users that further use the same email address, but change the
display name part. Since Chessboard is intended to work as a mailinglist
archive, it is important that each post is shown next to *exactly* the
display name it was made under. As a bonus, this commit adds an alias
tracking functionality that maps all encountered display names to the
user ID with the email address used to make the posts under, which
allows to display a list of known aliases on the user page for a user.