Automatically (un)subscribe users.
Users are first subscribed to a mailinglist when they first want to post
to it. This cannot be done on registration, because there might be
email-only users, and those should not be subscribed against their will
do another mailinglist the forum may cover, even if it's just the nomail
version.

On account deletion, however, the user is unsubscribed from all
mailinglists the forum covers, as that is closest to his intention most
likely.