Add syntax highlighting.
Determining the language from a snippet of code without further info is
pretty hard. To allow domain-specific heuristics, the programming
language detection heuristic has been outsourced into the configuration,
with a simple heuristic for Ruby/C++/XML being included by default.

The styling of the result is not optimal yet, but the basics work. What
is not a code snippet does not get highlighted, and what is code gets
highlighted. It even properly processes indented blocks, thanks to
Ruby's great StringScanner (strscan) library.