Process quotes into proper <blockquote> elements.
Mailinglist archive does not mean each and every space needs to be
preserved. It must just be taken care that the formatter is so decent
that it almost never makes a mistake when transforming an email into
HTML. The previous approach with parsing emails as markdown produced too
much errors in this regard.