Optimise thread and topic views' queries.
These views now don't query the database multiple times for each post in
a thread, but only once for all info in the entire thread. The topics
view is still a little unperformant, though, as it loads the entire
thread rather than just the displayed posts when it constructs the
information related to the whole thread. Still more performant than
before, though, as much less queries are made.