Don't crash on invalid post input and preserve form values.
The post creation routes do not save the Post instance do the database,
but instead send the post to the mailinglist. It's the mailinglist
handler which then saves it to the database. To catch user errors in
creation of posts, the bare database constraints thus are not enough, so
this commit employs explicit validations which can be checked
independantly of saving the post or not doing
that. Post#send_to_mailinglist explicitely queries the #valid? method
and if it fails, raises Sequel::ValidationFailed, mimicing the #save
method in its behaviour.

The :used_alias column's CHECK constraint is lowed to 1 character as
appearently some people like to send under a display name of "A", which
is only one character long. An empty display name (techinically that
would be »"" <email@address.invalid>« in the mail headers) is not handled
by this, but should cause the mail processor to use the email address as
the display name.