Rewrite Alias and LDAP systems.
LDAP integration does not work with my original conception of the alias
system. So this commit purges that one and instead gives each user a
display name as the source of the email's display name send in mails to
the mailinglist. Each user now has a UID, which is the user's canonical
nickname and is displayed in parentheses next to the post's used
alias (the display name found in the incoming email). That UID is used
for the mapping of the user to the LDAP server as well, thus a change is
not allowed.