Explicitely specify mail encoding.
The 'mail' gem tries to write a warning to the STDERR constant if you
don't do this. However, in the ml monitor daemon, STDERR is a closed IO
stream. As a result, the ml monitor crashes with an IOError exception,
which is swallowed as it is not catched anywhere and cannot be logged.

Writing on STDERR is nothing a library should do. It is a misbehaviour
in the mail gem, which should write to $stderr, which is properly
redirected.